+++ CHANGELOG.md
@@ -4,8 +4,17 @@ As this project is pre 1.0, breaking changes may happen for minor version bumps.
 
 ## Unreleased
 
+### Changes
 * Fix missing Liquidity Pool ID in AccountResponse Balance ([#379](https://github.com/stellar/java-stellar-sdk/pull/379)).
 * Fix null pointer when calling ChangeTrustOperationResponse.getAsset() for LiquidityPool trust line ([#378](https://github.com/stellar/java-stellar-sdk/pull/378)).
+
+### Breaking changes
+* Changed offer ids to be represented in requests and response models as long data type. ([#386](https://github.com/stellar/java-stellar-sdk/pull/386)).
+  * `TradesRequestBuilder.offerId()`
+  * `TradeResponse.getOfferId()` 
+  * `TradeResponse.getBaseOfferId()` 
+  * `TradeResponse.getCounterOfferId()` 
+  * `RevokeSponsorshipOperationResponse.getOfferId()`
 
 ## 0.29.0